Show sample of images on Dataset page
This also:
* factors a lot of the reactivesearch code, including the result
  card, out of reactive_search.js and into Components/search/
* gets rid of the datasetNames API lookup since dataset_name is now
  stored in ES
* marks one of the image thumbnails as the og:image, which fixes #389

This is not strictly in the current scope of works but I think is a
valuable UX improvement, as well as a code cleanup.
